+++ registry/default/ui/kbd.tsx
@@ -1,29 +1,22 @@
-'use client';
+import { cn } from '@/lib/utils';
 
-import * as React from 'react';
-import { cn } from '@/registry/default/lib/utils';
-import { cva, type VariantProps } from 'class-variance-authority';
-
-const kbdVariants = cva('inline-flex items-center justify-center font-mono rounded-md', {
-  variants: {
-    variant: {
-      default: 'bg-accent border border-border text-accent-foreground',
-      outline: 'text-accent-foreground border border-input',
-    },
-    size: {
-      md: 'h-7 min-w-7 px-1.5 text-xs [&_svg]:size-3.5',
-      sm: 'h-6 min-w-6 px-1 text-[0.75rem] leading-[0.75rem] [&_svg]:size-3',
-      xs: 'h-5 min-w-5 px-1 text-[0.6875rem] leading-[0.75rem] [&_svg]:size-3',
-    },
-  },
-  defaultVariants: {
-    variant: 'default',
-    size: 'md',
-  },
-});
-
-function Kbd({ className, variant, size, ...props }: React.ComponentProps<'kbd'> & VariantProps<typeof kbdVariants>) {
-  return <kbd data-slot="kbd" className={cn(kbdVariants({ variant, size }), className)} {...props} />;
+function Kbd({ className, ...props }: React.ComponentProps<'kbd'>) {
+  return (
+    <kbd
+      data-slot="kbd"
+      className={cn(
+        'bg-muted text-muted-foreground pointer-events-none inline-flex h-5 w-fit min-w-5 items-center justify-center gap-1 rounded-sm px-1 font-sans text-xs font-medium select-none',
+        "[&_svg:not([class*='size-'])]:size-3",
+        '[[data-slot=tooltip-content]_&]:bg-background/20 [[data-slot=tooltip-content]_&]:text-background dark:[[data-slot=tooltip-content]_&]:bg-background/10',
+        className,
+      )}
+      {...props}
+    />
+  );
 }
 
-export { Kbd, kbdVariants };
+function KbdGroup({ className, ...props }: React.ComponentProps<'div'>) {
+  return <kbd data-slot="kbd-group" className={cn('inline-flex items-center gap-1', className)} {...props} />;
+}
+
+export { Kbd, KbdGroup };