+++ registry/default/ui/select.tsx
@@ -52,8 +52,9 @@ const selectTriggerVariants = cva(
   {
     variants: {
       size: {
-        sm: 'h-7 px-2.5 text-xs gap-1 rounded-md',
-        md: 'h-8.5 px-3 text-[0.8125rem] leading-(--text-sm--line-height) gap-1 rounded-md',
+        xs: 'h-7 px-2 text-xs gap-1 rounded-md',
+        sm: 'h-8 px-2.5 text-xs gap-1 rounded-md',
+        md: 'h-9 px-3 text-sm gap-1 rounded-md',
         lg: 'h-10 px-4 text-sm gap-1.5 rounded-md',
       },
     },